pyproject.toml
[tool.poetry]
name = "repl-nix-databaseflask"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]
readme = "README.md"
packages = [{include = "repl_nix_databaseflask"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.2"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
